feat: implement declaration node parsing and codegen
- Add declaration_node type to AST with proper visitor pattern
- Implement parse_declaration() in parser to handle type declarations
- Add code generation for declarations with kind parameters
- Fix single_real_declaration test case by properly handling standalone declarations
- Update test expectations to match generated output

This addresses the user requirement that standalone declarations must be
properly handled in the AST pipeline, not dropped as empty literals.

+++ example/frontend_test_cases/multi_statement/multi_statement.f90
@@ -1,6 +1,6 @@
 program main
     implicit none
-    integer :: x
+    integer(4) :: x
     real(8) :: y
 
     x = 42
